add BC layer for changed return types
